Add files via upload
Web SQL is a deprecated web page API for storing or managing the data in databases that can be queried using a variant of SQL like creating databases, opening the transaction, creating tables, inserting values to tables, deleting values, and reading data. 

The Web SQL Database API is not a part of the HTML5 specification but it is a separate specification. It specifies a set of  APIs to manipulate the client-side databases using SQL. The web SQL database still works in Chromium-based browsers, but support is being phased out. Web SQL was deprecated and removed for third-party contexts in Chromium 97. Web SQL access in insecure contexts is deprecated as of Chromium 105 at which time a warning message will be shown in the Chrome DevTools Issue panel.

Methods:
There are three basic methods as shown below as follows:

Method	Action Performed
openDatabase	-It can be used to create a new database or can create the database object using an existing database.
transaction-It transaction can control a transaction and perform either commit or rollback depending upon the situation.
executeSql-It is used to execute a real SQL query.